Exclude non-existent ports from the switch "Port List"
Cisco Catlyst 6500 modular switches return ports on removed line cards
as if they were still present and there is no way to tell them apart
from real ports. To exclude them from view, Port List rows that
have module number (the first number in n/m designation) that does not
correspond to existing module are not displayed by view.cgi. Note,
that there are still entries for these interfaces in the database,
so this is more of a workaround than a proper fix.