Change the way Port Info UI is closed
Instead of closing upon click on the close button, use
custom event 'dismiss' on TD.pi-container and then let
'click' event on the button emit the 'dismiss' event.

This avoid the problem with no way to close the UI when
the button is not displayed (which can happen when the
button is temporarily hidden by the "are you sure?"
dialog.