Prepare ground for implementing semantic URLs
Parse the location.pathname and get "selector"
and "argument" out of it. Selector will be two letter
component selecting section of the application: 'sw'
for switch list/port list, 'sr' for search tool, 'ap'
for add patches, 'ab' for about. The additional
argument will then allow to select particular switch
etc.

For this to work, the web server needs to be configured
to strip away this additional information before
searching for actual files. Also, this requires the
document base to be explicitly set in the skeleton
index.html file through the BASE tag, otherwise
various external assets would be looked for in wrong
places; this base needs to be rewritten for deployment
or moving. This needs some way of being automated or
otherwise dealt with.