Implement internal state and central dispatch
In order to allow some minimal navigation using
the standard back and forward concepts of old,
we implemented the idea of central dispatch
(function shared.dispatch() defined in spam.js)
that invokes other parts of the aplication.
This happens upon user selecting an action (in
menu, via buttons) or by user navigating with
back/forward buttons (as if we weren't an SPA).
For this pushState/replaceState functions of the
History API are used.

Moreover, this brings in clean/semantic URLs, so
that sections of the app and some internal states
can be linked to (currently only the menu choices
and host selection within Switch List).