New config loading function
New function load_config_json(FILE) introduced. This
is the start of moving towards JSON based configuration.
This will have two benefits: 1) no config parsing needed
in the SPAMv2.pm library; 2) easy and flexible extending
of the config.

The JSON is parsed in relaxed mode to allow for
shell-like comments.

Note, that the in-memory structure of the config will
change so that all of the codebase needs to adapt.