Make getting ARP tables concurrent
This change adapts the code that runs concurrent host polls to be more
flexible and then adds getting arptable as another concurrent task.  So
with this change getting ARP will not run in the main process, but fork
another one that can run along the host polling processes, saving some
time.