Fix bug affecting duplex values saved to backend
There was stupid bug in sql_update_status(), that caused values using
portIndex as their table index to be stored incorrectly.  This only
affected prot duplex values.  Side effect of this was, that some ports
were always seen as changed since last check and therefore never aged
(their value in db backend and the one read from switch was different,
but broken saving never updated it in the db).