Fix some switches not being properly displayed in UI
Switches that happened to have non-alphanumeric characters in their name
were not handled correctly by the backend and queries for them were
empty.  This was caused by the fact that Port List uses the same query
that Search Tool uses.  Thus the parameters of the query were normalized,
which mangled their names.  Normalization of search query input is now
never normalized if 'mode' is 'portlist'.

This still leaves Search Tool unable to search for these switches.