Implement command line option --host, --hostre and --worklist
Semantics of --host option was changed so that it no longer does
start-anchored substring match.  New --hostre option accepts regexp that
selects hosts to be processed.  The two options' results are unified
together. Results of the --host and --hostre can be dry-tested with the
--worklist: it will display the list of hosts passing the filters
and then exit without doing anything else.