Add new section to poll_host() that uses snmp_get_object()
New section of the poll_host() uses previously implemented
snmp_get_object() function.  This section stores its output into
'mibs-new' subtree of $swdata to not conflict with the data from
snmp_get_tree().  Also, the indexing loops that create some auxiliary
indexes were converted to use the new structure.

This is the starting point for moving the rest of SPAM to the new
in-memory structure of SNMP data.  Once everything is migrated the old
section and snmp_get_tree() itself will be removed.