Fix Add Patches not displaying update summary any more
This was caused by introducing context helper for search by authname,
which caused exception when there was no 'fields' list in the 'search'
response as is case in the query response from adding new patches;
failing context helper caused dust.render to fail and therefore nothing
was shown to the user.  The context helper was updated to handle the
above case gracefully.