Add linecard_no() method to SPAM::Entity
This returns linecards's slot number.  This was previously done by
directly calling $entity->parent->entPhysicalParentRelPos, but now this
is the default of this method.  It is also possible to extract the slot
number from entity's entPhysicalName string using configurable regular
expression, as show in the example (the regex must match named capture
group 'slotno'):

  "entity-profiles": {
    "models": {
      "WS-C3850-48T": {
        "card_slot_no": "(Fixed|FRU Uplink) Module (?<slotno>\\d+)$",
      }
    }
  }