Implement additional way of extracting linecards from entities
Previously we were looking for a fixed pattern of
chassis-container-module in the entity tree.  Some switches, however
don't adhere to this, so we need alternative method of figuring out
what the linecards are in the full entity tree.  Therefore this commit
adds the ability to regex match entPhysicalName configurable per switch
model name (as extracted from $chassis->entPhysicalModelName).  For
example (Cisco Catalyst 3850):

  "entity-profiles": {
    "models": {
      "WS-C3850-48T": {
        "modules_by_name": "(Fixed|FRU Uplink) Module \\d+$",
      }
    }
  }