Improve message displaying machinery in some classes
We move the 'mesg' callback attribute to a reusable role, so we can equip
SPAM::Model::SNMP with it.  This allows for passing of the display
callback into deeper nested instances.  Ie.  PortToIfIndex wants to
display feedback message, but it's under SPAM::Model::SNMP, so this code
achieves that goal.