Convert webapp backend to Mojolicious
This takes the original spam-backend.cgi and turns it into a Legacy.pm
controller class that is referred to by router residing in Web.pm.  Note,
that this is just a superficial change -- most of the code is mostly
untouched.  Most importantly, the API between backend and frontend is
unchanged so that the old frontend still works.