Serve static files through Mojolicious
Previously we were relying on external (from the viewpoint of the
application) webserver to serve all static files.  This commit moves the
files so that they are served through the Mojolicious framework, and moves
the API endpoint under its own tree (/api/v0).  This will make the frontend
more flexible and allow for more convenient development.