Convert Misc::sql_site_uses_cp to Mojo::Pg
Also fix it as it was completely broken.  This function is a relict of a
bygone era where we had two-level patching hierarchy at some sites.  I think
at this moment there is only one left and folks there have not been
maintaining patching records for ages.  Maybe this whole thing should be
finally excised.