Switch configuration from JSON to perl
We have been using JSON as the format of the configuration file for long
time.  Unfortunately, JSON is not good format for this purpose and the
relaxed version with comments and trailing commas is not widely
supported.  This commit switches to using perl syntax configuration: The
file is basically a perl data structure using hashrefs and arrayrefs.